vader: add cawbird, mimeo, replace youtube-dl with yt-dlp
youtube-dl fails downloading from youtube at the moment
and also was rate-limited in the last months. yt-dlp is a fork
and worked much better.

mimeo is a nice mime helper to open files with the right application
(uses xdg), show what would be used to open (-c) or just the mime type
(-m).

cawbird is a GTK twitter client.