base-common: improve in/out, find_up command
`in` and `out` are now proper shell apps instead of zsh functions which
makes them easier to call (for use with `entr`, for example).
The commands now use the `find_up` command to find config files in parent
directories if the current doesn't have a config (like pytest with conftest.py)
`find_up` may also be useful as standalone command.